Remove LegacyDashletCard, refactor IconGuideDashlet #6247
