+++ ui/app/src/components/index.ts
@@ -70,6 +70,7 @@ export * from './DialogHeaderAction';
 export * from './DialogTitle';
 export * from './DraggablePanelListItem';
 export * from './DropDownMenuButton';
+export * from './DuplicateSiteDialog';
 export * from './EditFormPanel';
 export * from './EditGroupDialog';
 export * from './EditModeSwitch';
@@ -133,17 +134,6 @@ export * from './LegacyConcierge';
 export * from './LegacyDashboardFrame';
 export * from './LegacyFormDialog';
 export * from './LegacyIFrame';
-export * from './LegacySiteDashboard';
-export * from './LegacySiteDashboard/LegacyApprovedScheduledDashlet';
-export * from './LegacySiteDashboard/LegacyApprovedScheduledDashletGrid';
-export * from './LegacySiteDashboard/LegacyAwaitingApprovalDashlet';
-export * from './LegacySiteDashboard/LegacyAwaitingApprovalDashletGrid';
-export * from './LegacySiteDashboard/LegacyDashletCard';
-export * from './LegacySiteDashboard/LegacyInReviewDashlet';
-export * from './LegacySiteDashboard/LegacyRecentActivityDashlet';
-export * from './LegacySiteDashboard/LegacyRecentActivityDashletGrid';
-export * from './LegacySiteDashboard/LegacyRecentlyPublishedDashlet';
-export * from './LegacySiteDashboard/LegacyUnpublishedDashlet';
 export * from './LegacySiteToolsFrame';
 export * from './LoadingIconButton';
 export * from './LoadingState';
@@ -156,6 +146,7 @@ export * from './LoginForm';
 export * from './LoginView';
 export * from './LogoAndMenuBundleButton';
 export * from './MarketplaceDialog';
+export * from './MaxLengthCircularProgress';
 export * from './MediaCard';
 export * from './MinimizedBar';
 export * from './MinimizedBarPortal';
